requirements.txt
numpy
pandas
loguru
requests
langdetect
lxml
transformers
torch
scikit-learn
schedule
empyrical
nltk
jinja2
edgartools
openpyxl
openai
ollama
kaleido==0.1.0.post1
botrading==1.0.0